Fixxed Map
Made the map actually display the annotations

+++ VirtualTourist/VirtualTourist/Other/APICommands.swift
@@ -16,7 +16,7 @@ class APICommands{
    let methodHead = "/services/rest/?method="
    let getPhoto = "flickr.photos.search"
     //gets 30 photos at given lat+long
-    func getPhotos(lat: Float ,long: Float){
+    func getPhotos(lat: Double ,long: Double){
         let urlString = "\(webURL)\(methodHead)\(getPhoto)&api_key=\(apiKey)&accuracy=15&has_geo=1&lat=\(lat)&lon=\(long)&per_page=30&page=1&format=json&nojsoncallback=1"
         let url = URL(string: urlString)
         let request = URLRequest(url: url!)
